+++ BilleteraVirtualAPI/billeteravirtual/src/main/java/ar/com/ada/api/billeteravirtual/models/request/MovimientoRequest.java
@@ -1,11 +1,13 @@
 package ar.com.ada.api.billeteravirtual.models.request;
 
+import java.math.BigDecimal;
+
 public class MovimientoRequest {
 
-    public int billeteraId;
+    public Integer billeteraId;
     public String moneda;
     public String conceptoOperacion;
-    public double importe;
+    public BigDecimal importe;
     public String tipoOperacion;
     public String detalle;
